+++ src/main/java/de/samply/exporter/ExporterConst.java
@@ -242,5 +242,7 @@ public class ExporterConst {
     public final static String FHIR_SEARCH_LIST_PARAMETER = "_list";
     public final static String DEFAULT_GRAPH_FORMAT = "JSON";
     public final static String ERROR_MESSAGE_INTERRUPTED_QUERY_EXECUTION = "Query execution interrupted abruptly";
+    public final static String QUERY_CONTEXT_SEPARATOR = ";";
+    public final static String QUERY_CONTEXT_EQUAL = "=";
 
 }